Use object address for logging - fixes uniqueness problems

+++ include/vlc_messages.h
@@ -48,7 +48,7 @@
 typedef struct
 {
     int     i_type;                             /**< message type, see below */
-    int     i_object_id;
+    uintptr_t   i_object_id;
     const char *psz_object_type;
     char *  psz_module;
     char *  psz_msg;                            /**< the message itself */
